made hook useTodos
